Update how-to-get-rid-of-virus-error.md
Maybe adding a reminder to  make sure to change out the "your username" part in the "C:\Users\Your user name\AppData\Roaming\RaftModLoader" for your username, and not simply just copy and paste? I feel like this is a mistake often made in haste, which gets overlooked because you think you did everything the right way, and cant figure out what you possibly could have done wrong.

Also, maybe adding a tip to check the antivirus quarantine for files related to the launcher, if anything may have slipped through. I feel like HMLCore.exe gets mentinoned as an issue and sulotion many times in the discord support channel. Probably for the very reason people forget to change the username part for their own.  Maybe mentioning if u would find eg the HMLCore.exe in the quarantine, to just restore it and add exclusion to it. since this seems to be the biggest issue-file in the appdata/roaming/raftModLoader folder.